+++ cli/cli/commands/sandbox/sandbox.go
@@ -13,29 +13,16 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/best_effort_image_puller"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/enclave_manager"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/enclave_manager/enclave_context"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/execution_ids"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/logrus_log_levels"
-	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
-	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_consts"
-	"github.com/kurtosis-tech/kurtosis-core/commons/current_time_str_provider"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/repl_runner"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"golang.org/x/crypto/ssh/terminal"
-	"golang.org/x/sys/unix"
-	"google.golang.org/grpc"
-	"google.golang.org/protobuf/types/known/emptypb"
-	"io"
-	"net"
-	"os"
 	"strings"
 )
 
 const (
-	enclaveDataVolMountpointOnReplContainer = "/kurtosis-enclave-data"
-
-
 	shouldPublishPorts = true
 
 	isPartitioningEnabled = true
@@ -44,19 +31,6 @@ const (
 	javascriptReplImageArg = "javascript-repl-image"
 	kurtosisLogLevelArg = "kurtosis-log-level"
 
-	// This is the directory in which the node REPL is running inside the REPL container, which is where
-	//  we'll bind-mount the host machine's current directory into the container so the user can access
-	//  files on their host machine
-	workingDirpathInsideReplContainer = "/repl"
-
-	replContainerSuccessExitCode = 0
-
-	// WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING
-	// vvvvvvvvvvvvvvv If you change these, update the REPL Dockerfile!!! vvvvvvvvvvvv
-	replContainerKurtosisSocketEnvVar = "KURTOSIS_API_SOCKET"
-	replContainerEnclaveDataVolMountpointEnvVar = "ENCLAVE_DATA_VOLUME_MOUNTPOINT"
-	// ^^^^^^^^^^^^^^^ If you change these, update the REPL Dockerfile!!! ^^^^^^^^^^^^
-	// WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING WARNING
 )
 var defaultKurtosisLogLevel = logrus.InfoLevel.String()
 
@@ -121,11 +95,12 @@ func run(cmd *cobra.Command, args []string) error {
 
 	enclaveId := execution_ids.GetExecutionID()
 
-	enclaveManager := enclave_manager.NewEnclaveManager(dockerClient, apiContainerImage)
+	enclaveManager := enclave_manager.NewEnclaveManager(dockerClient)
 
 	enclaveCtx, err := enclaveManager.CreateEnclave(
 		context.Background(),
 		logrus.StandardLogger(),
+		apiContainerImage,
 		kurtosisLogLevel,
 		enclaveId,
 		isPartitioningEnabled,
@@ -147,151 +122,10 @@ func run(cmd *cobra.Command, args []string) error {
 	}()
 
 	logrus.Debug("Running REPL...")
-	if err := runReplContainer(dockerManager, enclaveCtx, jsReplImage); err != nil {
+	if err := repl_runner.RunREPL(enclaveCtx, jsReplImage); err != nil {
 		return stacktrace.Propagate(err, "An error occurred running the REPL container")
 	}
 	logrus.Debug("REPL exited")
-
-	return nil
-}
-
-func runReplContainer(
-	dockerManager *docker_manager.DockerManager,
-	enclaveCtx *enclave_context.EnclaveContext,
-	javascriptReplImage string,
-) error {
-	enclaveId := enclaveCtx.GetEnclaveID()
-	networkId := enclaveCtx.GetNetworkID()
-	apiContainerHostMachinePortBinding := enclaveCtx.GetAPIContainerHostPortBinding()
-	kurtosisApiContainerIpAddr := enclaveCtx.GetAPIContainerIPAddr()
-	enclaveObjNameProvider := enclaveCtx.GetObjectNameProvider()
-	enclaveObjLabelsProvider := enclaveCtx.GetObjectLabelsProvider()
-
-	apiContainerUrlOnHostMachine := fmt.Sprintf(
-		"%v:%v",
-		apiContainerHostMachinePortBinding.HostIP,
-		apiContainerHostMachinePortBinding.HostPort,
-	)
-	conn, err := grpc.Dial(apiContainerUrlOnHostMachine, grpc.WithInsecure())
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred dialling the API container via its host machine port binding")
-	}
-	defer conn.Close()
-	apiContainerClient := kurtosis_core_rpc_api_bindings.NewApiContainerServiceClient(conn)
-
-	startRegistrationResp, err := apiContainerClient.StartExternalContainerRegistration(context.Background(), &emptypb.Empty{})
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred starting the registration of the interactive REPL container")
-	}
-	replContainerRegistrationKey := startRegistrationResp.RegistrationKey
-	replContainerIpAddrStr := startRegistrationResp.IpAddr
-	replContainerIpAddr := net.ParseIP(replContainerIpAddrStr)
-	if replContainerIpAddr == nil {
-		return stacktrace.NewError(
-			"Received an IP, '%v', from the API container to give the interactive REPL container, but it wasn't parseable to an IP",
-			replContainerIpAddr,
-		)
-	}
-
-	stdoutFd := int(os.Stdout.Fd())
-	windowSize, err := unix.IoctlGetWinsize(stdoutFd, unix.TIOCGWINSZ)
-	if err != nil {
-		return stacktrace.NewError("An error occurred getting the current terminal window size")
-	}
-	interactiveModeTtySize := &docker_manager.InteractiveModeTtySize{
-		Height: uint(windowSize.Row),
-		Width:  uint(windowSize.Col),
-	}
-
-	// Map of host_path -> path_on_container that will mounted
-	var bindMounts map[string]string
-	hostMachineWorkingDirpath, err := os.Getwd()
-	if err != nil {
-		logrus.Warn("Couldn't get the current working directory; local files will not be available")
-		bindMounts = map[string]string{}
-	} else {
-		bindMounts = map[string]string{
-			hostMachineWorkingDirpath: workingDirpathInsideReplContainer,
-		}
-	}
-
-	interactiveReplGuid := current_time_str_provider.GetCurrentTimeStr()
-
-
-	kurtosisApiContainerSocket := fmt.Sprintf("%v:%v", kurtosisApiContainerIpAddr, kurtosis_core_rpc_api_consts.ListenPort)
-	containerName := enclaveObjNameProvider.ForInteractiveREPLContainer(interactiveReplGuid)
-	labels := enclaveObjLabelsProvider.ForInteractiveREPLContainer(interactiveReplGuid)
-	// TODO Add interactive labels!!!
-	createAndStartArgs := docker_manager.NewCreateAndStartContainerArgsBuilder(
-		javascriptReplImage,
-		containerName,
-		networkId,
-	).WithInteractiveModeTtySize(
-		interactiveModeTtySize,
-	).WithStaticIP(
-		replContainerIpAddr,
-	).WithEnvironmentVariables(map[string]string{
-		replContainerKurtosisSocketEnvVar:            kurtosisApiContainerSocket,
-		replContainerEnclaveDataVolMountpointEnvVar: enclaveDataVolMountpointOnReplContainer,
-	}).WithBindMounts(
-		bindMounts,
-	).WithVolumeMounts(map[string]string{
-		enclaveId: enclaveDataVolMountpointOnReplContainer,
-	}).WithLabels(
-		labels,
-	).Build()
-	replContainerId, _, err := dockerManager.CreateAndStartContainer(context.Background(), createAndStartArgs)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred starting the REPL container")
-	}
-	defer func() {
-		// Safeguard to ensure we don't leak a container
-		if err := dockerManager.KillContainer(context.Background(), replContainerId); err != nil {
-			logrus.Errorf("An error occurred killing the REPL container:")
-			fmt.Fprintln(logrus.StandardLogger().Out, err)
-		}
-	}()
-
-	finishRegistrationArgs := &kurtosis_core_rpc_api_bindings.FinishExternalContainerRegistrationArgs{
-		RegistrationKey: replContainerRegistrationKey,
-		ContainerId:     replContainerId,
-	}
-	if _, err := apiContainerClient.FinishExternalContainerRegistration(context.Background(), finishRegistrationArgs); err != nil {
-		return stacktrace.Propagate(err, "An error occurred finishing the registration of the interactive REPL container")
-	}
-
-	hijackedResponse, err := dockerManager.AttachToContainer(context.Background(), replContainerId)
-	if err != nil {
-		return stacktrace.Propagate(err, "Couldn't attack to the REPL container")
-	}
-	defer hijackedResponse.Close()
-
-	// From this point on down, I don't know why it works.... but it does
-	// I just followed the solution here: https://stackoverflow.com/questions/58732588/accept-user-input-os-stdin-to-container-using-golang-docker-sdk-interactive-co
-	go io.Copy(os.Stderr, hijackedResponse.Reader)
-	go io.Copy(os.Stdout, hijackedResponse.Reader)
-	go io.Copy(hijackedResponse.Conn, os.Stdin)
-
-	stdinFd := int(os.Stdin.Fd())
-	var oldState *terminal.State
-	if terminal.IsTerminal(stdinFd) {
-		oldState, err = terminal.MakeRaw(stdinFd)
-		if err != nil {
-			// print error
-			return stacktrace.Propagate(err, "An error occurred making STDIN stream raw")
-		}
-		defer terminal.Restore(stdinFd, oldState)
-	}
-
-	exitCode, err := dockerManager.WaitForExit(context.Background(), replContainerId)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred waiting for the REPL container to exit")
-	}
-	if exitCode != replContainerSuccessExitCode {
-		logrus.Warnf("The REPL container exited with a non-%v exit code '%v'", replContainerSuccessExitCode, exitCode)
-	}
-
-	terminal.Restore(stdinFd, oldState)
 
 	return nil
 }